% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{imvigor_mutations}
\alias{imvigor_mutations}
\title{Binary somatic mutation indices in TCGA BLCA urothelial cancers.}
\format{
A data frame with 180 rows (cancer samples) and 394 variables:
\itemize{
\item \strong{sample_id} contains unique sample identifiers
\item \strong{clust_id} codes for molecular cluster assignment
\item the remaining columns are named after HUGO gene symbols and store binarized
indexes of mutation indices. 0 represents the wild-type variant, 1 represents
at least one somatic mutation in the given gene.
}
}
\source{
IMvigor cohort mutation data set authored by Si Yangming and others
(DOI: 10.1038/nature25501) and available as CC 3.0 licensed R package
\href{https://github.com/SiYangming/IMvigor210CoreBiologies}{IMvigor210CoreBiologies}.
}
\usage{
data(imvigor_mutations)
}
\description{
A data frame with binary indexes of mutations in over 300 genes in a set of
urothelial cancers assigned to three molecular clusters.
}
