% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_df}
\alias{process_df}
\title{Pre-processing of a data frame for permutation testing.}
\usage{
process_df(x)
}
\arguments{
\item{x}{a data frame.}
}
\value{
a 0/1 matrix.
}
\description{
Throws errors if any any of the columns is in an unsupported
non-binary format. Converts two-level factors to 0/1 integers.
}
