% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_colnames_}
\alias{set_colnames_}
\title{Set column names.}
\usage{
set_colnames_(x, nm)
}
\arguments{
\item{x}{a data frame or a matrix.}

\item{nm}{a character vector with column names.}
}
\value{
a matrix or a data frame.
}
\description{
Sets column names in a matrix or a data frame.
}
