% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_vector}
\alias{split_vector}
\title{Split a vector into a list of its fragments.}
\usage{
split_vector(x, n)
}
\arguments{
\item{x}{a vector.}

\item{n}{number of chunks.}
}
\value{
a list of vectors
}
\description{
The function takes a vector and returns a list of its chunks of possibly
similar sizes.
}
