% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{tcga_mutations}
\alias{tcga_mutations}
\title{Binary somatic mutation indices in TCGA BLCA urothelial cancers.}
\format{
A data frame with 405 rows (cancer samples) and 17005 variables:
\itemize{
\item \strong{sample_id} contains unique sample identifiers
\item \strong{clust_id} codes for molecular cluster assignment
\item the remaining columns are named after HUGO gene symbols and store binarized
indexes of mutation indices. 0 represents the wild-type variant, 1 represents
at least one somatic mutation in the given gene.
}
}
\source{
TCGA BLCA mutation data set generated with the text file deposited at
\href{https://www.cbioportal.org/study/summary?id=blca_tcga_pan_can_atlas_2018}{cBioportal}.
}
\usage{
data(tcga_mutations)
}
\description{
A data frame with binary indexes of mutations in over 17K genes in a set of
urothelial cancers assigned to three molecular clusters.
}
